Moved logging related logic to separate file.
